# Conflicts:
#	core/src/main/java/org/seismic/arcade/Controller.java
#	core/src/main/java/org/seismic/arcade/Model.java

+++ core/src/main/java/org/seismic/arcade/Model.java
@@ -20,7 +20,7 @@ public class Model {
     public Vector2 vel;
     public Vector2 force;
     public int r;
-    public double angle;
+    public float angle;
     public boolean boost;
     public boolean landed;
     public boolean crashed;
